+++ src/main/kotlin/me/bossm0n5t3r/leetcode/firstUniqueCharacterInAString/FirstUniqueCharacterInAString.kt
@@ -1,0 +1,17 @@
+package me.bossm0n5t3r.leetcode.firstUniqueCharacterInAString
+
+class FirstUniqueCharacterInAString {
+    class Solution {
+        fun firstUniqChar(s: String): Int {
+            val result = mutableMapOf<Char, MutableList<Int>>()
+            for (i in s.indices) {
+                if (result.containsKey(s[i])) {
+                    result[s[i]]!!.add(i)
+                } else {
+                    result[s[i]] = mutableListOf(i)
+                }
+            }
+            return result.values.filter { it.size < 2 }.map { it[0] }.minOrNull() ?: -1
+        }
+    }
+}
